refactor(core): split governance_handlers — extract treasury module
Extract treasury governance handlers from governance_handlers/mod.rs
into a dedicated treasury.rs submodule. Pure structural surgery with
no behavioral changes.

What moved to treasury.rs (2024 lines):
- validate_positive_amount() and validate_currency_match() helpers
- handle_treasury_proposal() dispatcher
- handle_treasury_create_budget(), handle_treasury_withdrawal()
- handle_treasury_modify_spending_rule()
- handle_treasury_transfer_between_budgets()
- handle_treasury_cancel_budget(), handle_treasury_reclaim_budget()
- handle_budget_proposal() (core ledger handler)
- execute_surplus_allocation(), execute_share_redemption()
- execute_bond_issuance()

File sizes after split:
- mod.rs: 1500 lines (was 3503)
- treasury.rs: 2024 lines
- federation.rs: 1119 lines (unchanged)
